[optimise] Split BTree cache into levels to avoid busy loop
Fix performance bug in BTreeCache that profiling picks up; as soon as it is full, and because it favours inner nodes, `removeNext()` ends up traversing a cache full of inner nodes looking for an outer to remove, before giving up and removing an inner. The solution is to have separate inner and outer node caches.

This