[optimize] Make Paged use MappedByteBuffer
We get better performance reading (and writing probably isn’t worse) when we use MappedByteBuffer instead of RandomAccessFile.

How do we handle the case where we are writing beyonf the end of file ?

We start by not allowing it and throwing an exception.
Now we need to wrap the writing in some kind of helper class, and flip to using RandomAccessFile when we need to extend the file.