Paged - s/RandomAccessFile/MappedByteBuffer/g
This is a VERY quick hack to test whether we get performance by mapping paged files instead of using an RAF; we do, so it stays for now, but there are a lot of questions. Does write() even work correctly yet ?

An alternative may be to use the org.apache.pdfbox.io.RandomAccessFile as a drop in replacement offering potential performance gain.

But yes, the big CPU hogs have disappeared, and we seem to be at low 23s rather than 24s; costs gone into some background threads somewhere, I think ?

Thorw an error if we try to write; just until we are sure it works.