Paged - s/RandomAccessFile/MappedByteBuffer/g
This is a VERY quick hack to test whether we get performance by mapping paged files instead of using an RAF; we do, so it stays for now, but there are a lot of questions. Does write() even work correctly yet ?

An alternative may be to use the org.apache.pdfbox.io.RandomAccessFile as a drop in replacement offering potential performance gain.

But yes, goes from about 7.7s to about 7.2s, and profiles agree that the big CPU hogs have disappeared.