Increased locking granularity
Drastic locking fix; NodeIterator locks and unlocks twice (the twice is a bug, IMHO) on each hasNext(); next() - if we hold the lock for the duration of the iteration instead, we save 20% of performance time. That’s nuts.

40 repeats took time from 59.5 s to 47s, almost entirely in
Serializer.serializeToReceiver, from 25s to 12s.

We had experimented with a fix that replaces hasNext();next() by relying on next() returning null and not throw()-ing - that got about half the speedup, logically enough.

This one is better, and should be productized.

NOTE also - tried to find if RawNodeIterator was amenable to the same fix, but removing locks there made little difference. I guess that’s just a case of the workload being different, but we can move on.