[optimize] Use FileChannel for extending IO
When we open a file via a MappedByteBuffer, then extend it, we need to remap; for now we just read/write beyonf the end of the mapped region with the FileChannel. We should get the whole file mapped the next time we open/close it.

We could look at remapping (or adding an extra mapping) if the file has been significantly written since it was opened.