Cache index query results
Initially a global range cache holding the result set for an index lookup which is repeated repeatedly.
TODO (AP) - we want to restrict cacheing to the context of a single query. Otherwise the cache will grow; if we attach it to the query root, we can at least GC it.