[bugfix] restore optimizable to v6 behaviour
The PR https://github.com/eXist-db/exist/pull/4869 introduced subtly different optimization behaviour which differs to that of V6.3 on a specific large workload; the V6.3 behaviour appears to be correct.

This change re-instates the V6 behaviour, by explicitly distinguishing between a non-optimizable sequence and a (possibly optimizable) non-empty sequence.

TBD (AP) we do not have a tractable test for changed behaviour, it is all horribly subtle, and we need to avoid a regression. This one took 2 years to manifest.