[refactor] Introduce Prop class (properties)
Abstract constant configuration items to be a `Prop` object that has both an interned `Str` and the raw String, to be used according to which type is appropriate. Prefer the `Str` part wherever we can, because it has fast lookup in hash tables, which appears to be key to performance.